+++ bitzel/Cargo.toml
@@ -20,6 +20,7 @@ tracing = { workspace = true }
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 walkdir = "2.5"
 num_cpus = "1.16"  # For detecting CPU count in CLI configuration
+clap = { version = "4.5", features = ["derive"] }
 
 # Sandboxing with Linux namespaces
 nix = { version = "0.29", features = ["mount", "sched", "process", "user", "fs"] }
