+++ bitzel/src/main.rs
@@ -1,29 +1,13 @@
 //! Bitzel - Bazel-inspired build orchestrator for BitBake/Yocto projects
 //!
-//! Orchestrates:
-//! 1. KAS configuration loading
-//! 2. Repository fetching
-//! 3. Configuration file generation (local.conf, bblayers.conf)
-//! 4. Layer context with priorities and overrides
-//! 5. Recipe parsing with BuildContext (respects bbappends, layer priorities)
-//! 6. Dependency graph building (using RecipeGraph)
-//! 7. Task graph construction (using TaskGraphBuilder)
-//! 8. Cached execution (using AsyncTaskExecutor)
+//! Supports two modes of operation:
+//! 1. KAS mode: Build using KAS configuration files
+//! 2. Native BitBake mode: Build using standard BitBake configuration
 
-use convenient_bitbake::{
-    BuildContext, ExtractionConfig, RecipeExtractor, RecipeGraph,
-    TaskGraphBuilder, InteractiveExecutor, InteractiveOptions, TaskSpec,
-    TaskExtractor, TaskImplementation,
-    Pipeline, PipelineConfig,
-    SignatureCache, SignatureStats, EnhancedTaskSignature,
-};
-use convenient_bitbake::executor::types::{NetworkPolicy, ResourceLimits};
-use convenient_kas::{ConfigGenerator, include_graph::KasIncludeGraph};
-use std::collections::HashMap;
-use std::fs;
-use std::path::{Path, PathBuf};
-use std::process::Command;
-use std::time::Duration;
+mod commands;
+
+use clap::Parser;
+use commands::{Cli, Commands};
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
 #[tokio::main]
@@ -37,544 +21,30 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
         .with(tracing_subscriber::fmt::layer())
         .init();
 
+    // Parse command-line arguments
+    let cli = Cli::parse();
+
     println!("\n╔════════════════════════════════════════════════════════╗");
     println!("║              BITZEL BUILD ORCHESTRATOR                 ║");
     println!("║  Layer-aware BitBake with override resolution         ║");
     println!("╚════════════════════════════════════════════════════════╝\n");
 
-    // Configuration
-    let kas_file = Path::new("examples/busybox-qemux86-64.yml");
-    let build_dir = PathBuf::from("build");
-    let repos_dir = build_dir.join("repos");
-    let conf_dir = build_dir.join("conf");
-
-    if !kas_file.exists() {
-        eprintln!("❌ Kas file not found: {}", kas_file.display());
-        eprintln!("   Run this from the repository root directory");
-        std::process::exit(1);
-    }
-
-    fs::create_dir_all(&build_dir)?;
-
-    // ========== Step 1: Load KAS Configuration ==========
-    tracing::info!("Loading kas configuration from {}", kas_file.display());
-    let kas_config_graph = KasIncludeGraph::build(kas_file).await?;
-    let kas_config = kas_config_graph.merge_config();
-
-    println!("📋 Configuration:");
-    println!("  Machine:  {}", kas_config.machine.as_deref().unwrap_or("unknown"));
-    println!("  Distro:   {}", kas_config.distro.as_deref().unwrap_or("unknown"));
-    if let Some(targets) = &kas_config.target {
-        println!("  Targets:  {}", targets.join(", "));
-    }
-    println!();
-
-    // ========== Step 2: Fetch Repositories ==========
-    println!("📦 Fetching repositories...");
-    fs::create_dir_all(&repos_dir)?;
-
-    let mut layer_paths: HashMap<String, Vec<PathBuf>> = HashMap::new();
-
-    for (repo_name, repo) in &kas_config.repos {
-        if let Some(url) = &repo.url {
-            let repo_dir = repos_dir.join(repo_name);
-
-            if repo_dir.exists() {
-                tracing::debug!("Repository {} already exists", repo_name);
-            } else {
-                let branch = repo.branch.as_deref().unwrap_or("master");
-                tracing::info!("Cloning {} (branch: {})...", repo_name, branch);
-
-                let output = Command::new("git")
-                    .args(&[
-                        "clone",
-                        "--branch",
-                        branch,
-                        "--depth",
-                        "1",
-                        url,
-                        repo_dir.to_str().unwrap(),
-                    ])
-                    .output()?;
-
-                if !output.status.success() {
-                    tracing::error!(
-                        "Failed to clone {}: {}",
-                        repo_name,
-                        String::from_utf8_lossy(&output.stderr)
-                    );
-                    continue;
-                }
-            }
-
-            // Collect layer paths for this repo
-            let mut repo_layers = Vec::new();
-            for (layer_name, _) in &repo.layers {
-                let layer_path = repo_dir.join(layer_name);
-                if layer_path.exists() {
-                    repo_layers.push(layer_path);
-                }
-            }
-            layer_paths.insert(repo_name.clone(), repo_layers);
+    // Dispatch to appropriate command
+    match cli.command {
+        Commands::Kas { config, builddir, target } => {
+            println!("Mode: KAS Configuration");
+            println!("Config: {:?}", config);
+            println!();
+            commands::kas::execute(&config, &builddir, target).await?;
+        }
+        Commands::Build { builddir, target } => {
+            println!("Mode: Native BitBake");
+            println!("Build directory: {:?}", builddir);
+            println!("Target: {}", target);
+            println!();
+            commands::build::execute(&builddir, &target).await?;
         }
     }
-
-    let total_layers: usize = layer_paths.values().map(|v| v.len()).sum();
-    println!("  Cloned {} repos with {} layers", layer_paths.len(), total_layers);
-    println!();
-
-    // ========== Step 3: Generate BitBake Configuration Files ==========
-    println!("📝 Generating BitBake configuration...");
-    fs::create_dir_all(&conf_dir)?;
-
-    let config_gen = ConfigGenerator::new(&build_dir, kas_config.clone(), layer_paths.clone());
-    config_gen.generate_all().await?;
-
-    println!("  Generated: conf/local.conf");
-    println!("  Generated: conf/bblayers.conf");
-    println!();
-
-    // ========== Step 4: Build Layer Context with Priorities ==========
-    println!("🏗️  Building layer context with priorities...");
-
-    let mut build_context = BuildContext::new();
-
-    // Set machine and distro for override resolution
-    if let Some(machine) = &kas_config.machine {
-        build_context.set_machine(machine.clone());
-    }
-    if let Some(distro) = &kas_config.distro {
-        build_context.set_distro(distro.clone());
-    }
-
-    // Add layers with their priorities from layer.conf
-    for (_repo_name, layers) in &layer_paths {
-        for layer_path in layers {
-            let layer_conf = layer_path.join("conf/layer.conf");
-            if layer_conf.exists() {
-                match build_context.add_layer_from_conf(&layer_conf) {
-                    Ok(()) => {}
-                    Err(e) => {
-                        tracing::warn!("Failed to parse layer.conf for {}: {}", layer_path.display(), e);
-                        // Create a default layer config
-                        let default_layer = convenient_bitbake::layer_context::LayerConfig {
-                            layer_dir: layer_path.clone(),
-                            collection: layer_path.file_name()
-                                .and_then(|s| s.to_str())
-                                .unwrap_or("unknown")
-                                .to_string(),
-                            priority: 5, // Default priority
-                            version: None,
-                            depends: vec![],
-                            series_compat: vec![],
-                            variables: HashMap::new(),
-                        };
-                        build_context.add_layer(default_layer);
-                    }
-                }
-            }
-        }
-    }
-
-    println!("  Loaded {} layers with priorities", build_context.layers.len());
-    for layer in &build_context.layers {
-        println!("    • {} (priority: {})", layer.collection, layer.priority);
-    }
-    println!();
-
-    // ========== Step 5: Parse Recipes with Parallel Pipeline ==========
-    println!("🔍 Parsing BitBake recipes with parallel pipeline...");
-
-    // Configure parallel pipeline
-    let pipeline_config = PipelineConfig {
-        max_io_parallelism: 32,
-        max_cpu_parallelism: num_cpus::get(),
-        enable_cache: true,
-        cache_dir: build_dir.join("bitzel-cache/pipeline"),
-    };
-
-    println!("  Parallelism: {} I/O tasks, {} CPU cores",
-             pipeline_config.max_io_parallelism,
-             pipeline_config.max_cpu_parallelism);
-
-    let pipeline = Pipeline::new(pipeline_config, build_context);
-
-    // Stage 1: Discover recipes in parallel
-    let (recipe_files, discover_hash) = pipeline.discover_recipes(&layer_paths).await?;
-    pipeline.save_stage_hash(&discover_hash).await?;
-
-    // Stage 2: Parse recipes in parallel
-    let (parsed_recipes, parse_hash) = pipeline.parse_recipes(recipe_files).await?;
-    pipeline.save_stage_hash(&parse_hash).await?;
-
-    // Extract task implementations from parsed recipes
-    let mut recipe_task_impls: HashMap<String, HashMap<String, TaskImplementation>> = HashMap::new();
-    for parsed in &parsed_recipes {
-        if !parsed.task_impls.is_empty() {
-            recipe_task_impls.insert(parsed.file.name.clone(), parsed.task_impls.clone());
-        }
-    }
-
-    println!("  Extracted {} task implementations from {} recipes",
-             recipe_task_impls.values().map(|m| m.len()).sum::<usize>(),
-             recipe_task_impls.len());
-    println!();
-
-    // ========== Step 6: Build Recipe Graph ==========
-    println!("🔗 Building recipe dependency graph...");
-
-    let mut extraction_config = ExtractionConfig::default();
-    extraction_config.extract_tasks = true;
-    extraction_config.resolve_providers = true;
-    extraction_config.use_python_executor = false;
-
-    let extractor = RecipeExtractor::new(extraction_config);
-
-    // Stage 3: Build recipe graph (sequential - needs all recipes)
-    let (graph, graph_hash) = pipeline.build_recipe_graph(&parsed_recipes, &extractor)?;
-    pipeline.save_stage_hash(&graph_hash).await?;
-
-    let stats = graph.statistics();
-    println!("  Recipes:       {}", stats.recipe_count);
-    println!("  Tasks:         {}", stats.task_count);
-    println!("  Dependencies:  {}", stats.total_dependencies);
-    println!("  Providers:     {}", stats.provider_count);
-    println!();
-
-    // ========== Step 7: Build Execution Plan ==========
-    println!("📊 Building execution plan...");
-
-    // Check for circular dependencies (warn but continue)
-    let build_order = match graph.topological_sort() {
-        Ok(order) => {
-            println!("  Build order: {} recipes", order.len());
-            order
-        }
-        Err(_e) => {
-            eprintln!("⚠️  Warning: Circular dependencies detected!");
-            let cycles = graph.detect_cycles();
-            if !cycles.is_empty() {
-                eprintln!("\nCircular dependency cycles found:");
-                for cycle in cycles.iter().take(5) {
-                    let names: Vec<String> = cycle.iter()
-                        .filter_map(|id| graph.get_recipe(*id).map(|r| r.name.clone()))
-                        .collect();
-                    if !names.is_empty() {
-                        eprintln!("  • {}", names.join(" → "));
-                    }
-                }
-                eprintln!("\n💡 Continuing with partial build (TaskGraphBuilder will handle cycles)...\n");
-            }
-            // Return empty order, let TaskGraphBuilder handle it
-            Vec::new()
-        }
-    };
-
-    // Show build plan for target recipes
-    if let Some(targets) = &kas_config.target {
-        println!("\n📋 Target recipes and their dependencies:");
-        for target in targets {
-            if let Some(recipe_id) = graph.find_recipe(target) {
-                let deps = graph.get_all_dependencies(recipe_id);
-                println!("\n  {} requires {} dependencies:", target, deps.len());
-
-                for dep_id in deps.iter().take(10) {
-                    if let Some(dep_recipe) = graph.get_recipe(*dep_id) {
-                        println!("    • {}", dep_recipe.name);
-                    }
-                }
-                if deps.len() > 10 {
-                    println!("    ... and {} more", deps.len() - 10);
-                }
-            } else {
-                println!("  ⚠️  {} not found in layers", target);
-            }
-        }
-    }
-
-    println!("\n✅ Build plan complete!");
-    println!();
-
-    // ========== Step 8: Build Task Execution Graph ==========
-    println!("🔨 Building task execution graph...");
-
-    let task_builder = TaskGraphBuilder::new(graph);
-    let task_graph = match task_builder.build_full_graph() {
-        Ok(tg) => tg,
-        Err(e) => {
-            eprintln!("❌ Failed to build task graph: {}", e);
-            return Err(e.into());
-        }
-    };
-
-    let stats = task_graph.stats();
-    println!("  Total tasks:   {}", stats.total_tasks);
-    println!("  Root tasks:    {}", stats.root_tasks);
-    println!("  Leaf tasks:    {}", stats.leaf_tasks);
-    println!("  Max depth:     {}", stats.max_depth);
-    println!();
-
-    // ========== Step 8.5: Compute Task Signatures ==========
-    println!("🔑 Computing task signatures (content-addressable cache)...");
-
-    // Build recipe hash map from parsed recipes
-    let mut recipe_hashes = HashMap::new();
-    for parsed in &parsed_recipes {
-        recipe_hashes.insert(parsed.file.name.clone(), parsed.hash.clone());
-    }
-
-    // Initialize signature cache
-    let mut sig_cache = SignatureCache::new(build_dir.join("bitzel-cache/signatures"));
-
-    // Load previous signatures for comparison BEFORE computing new ones
-    let cache_path = build_dir.join("bitzel-cache/signatures/signatures.json");
-    let previous_signatures = if cache_path.exists() {
-        let json = tokio::fs::read_to_string(&cache_path).await?;
-        serde_json::from_str::<HashMap<String, EnhancedTaskSignature>>(&json)
-            .unwrap_or_default()
-    } else {
-        HashMap::new()
-    };
-
-    // Compute NEW signatures for all tasks (don't load old ones first)
-    sig_cache.compute_signatures(
-        &task_graph,
-        &recipe_hashes,
-        &recipe_task_impls,
-        kas_config.machine.as_deref(),
-        kas_config.distro.as_deref(),
-    ).await?;
-
-    // ========== Step 8.6: Analyze Incremental Build Requirements ==========
-    println!("📊 Analyzing incremental build requirements...");
-
-    // Analyze which tasks need rebuilding
-    let mut tasks_to_rebuild = Vec::new();
-    let mut tasks_unchanged = Vec::new();
-    let mut new_tasks = Vec::new();
-
-    for (_task_id, task) in &task_graph.tasks {
-        let task_key = format!("{}:{}", task.recipe_name, task.task_name);
-
-        // Get current signature
-        if let Some(current_sig_str) = sig_cache.get_signature(&task.recipe_name, &task.task_name) {
-            // Check if we have a previous signature
-            if let Some(prev_sig) = previous_signatures.get(&task_key) {
-                if let Some(prev_sig_str) = &prev_sig.signature {
-                    if current_sig_str == prev_sig_str {
-                        tasks_unchanged.push(task_key.clone());
-                    } else {
-                        tasks_to_rebuild.push((task_key.clone(), "signature changed".to_string()));
-                    }
-                } else {
-                    tasks_to_rebuild.push((task_key.clone(), "no previous signature".to_string()));
-                }
-            } else {
-                new_tasks.push(task_key.clone());
-            }
-        }
-    }
-
-    // Save updated signatures
-    sig_cache.save().await?;
-
-    // Report statistics
-    let total_tasks = task_graph.tasks.len();
-    println!("  Total tasks:       {}", total_tasks);
-    println!("  Unchanged:         {} ({:.1}%)",
-        tasks_unchanged.len(),
-        (tasks_unchanged.len() as f64 / total_tasks as f64) * 100.0
-    );
-    println!("  Need rebuild:      {} ({:.1}%)",
-        tasks_to_rebuild.len(),
-        (tasks_to_rebuild.len() as f64 / total_tasks as f64) * 100.0
-    );
-    println!("  New tasks:         {} ({:.1}%)",
-        new_tasks.len(),
-        (new_tasks.len() as f64 / total_tasks as f64) * 100.0
-    );
-
-    // Show reasons for rebuild (first 10)
-    if !tasks_to_rebuild.is_empty() {
-        println!("\n  Sample rebuild reasons:");
-        for (task_key, reason) in tasks_to_rebuild.iter().take(10) {
-            println!("    • {}: {}", task_key, reason);
-        }
-        if tasks_to_rebuild.len() > 10 {
-            println!("    ... and {} more", tasks_to_rebuild.len() - 10);
-        }
-    }
-
-    println!();
-
-    // ========== Step 9: Create Task Specifications ==========
-    println!("📝 Creating task specifications with real implementations...");
-
-    let mut task_specs = HashMap::new();
-    let mut tasks_with_impl = 0;
-    let mut tasks_with_placeholder = 0;
-    let tmp_dir = build_dir.join("tmp");
-    fs::create_dir_all(&tmp_dir)?;
-
-    for (_task_id, task) in &task_graph.tasks {
-        let task_key = format!("{}:{}", task.recipe_name, task.task_name);
-
-        // Try to find real task implementation
-        let script = if let Some(recipe_impls) = recipe_task_impls.get(&task.recipe_name) {
-            if let Some(task_impl) = recipe_impls.get(&task.task_name) {
-                tasks_with_impl += 1;
-
-                // Use real implementation
-                let mut full_script = String::new();
-
-                // Add BitBake environment variables that tasks expect
-                full_script.push_str("set -e\n");
-                full_script.push_str("# BitBake environment\n");
-                full_script.push_str("export PN=\"${PN:-");
-                full_script.push_str(&task.recipe_name);
-                full_script.push_str("}\"\n");
-                full_script.push_str("export WORKDIR=\"${WORKDIR:-/work}\"\n");
-                full_script.push_str("export S=\"${S:-${WORKDIR}/src}\"\n");
-                full_script.push_str("export B=\"${B:-${WORKDIR}/build}\"\n");
-                full_script.push_str("export D=\"${D:-${WORKDIR}/outputs}\"\n\n");
-
-                // Add the actual task implementation
-                full_script.push_str(&task_impl.code);
-                full_script.push_str("\n\n");
-
-                // Create output marker for verification
-                let output_file = format!("{}.done", task.task_name);
-                full_script.push_str(&format!(
-                    "# Mark task complete\nmkdir -p /work/outputs\necho 'completed' > /work/outputs/{}\n",
-                    output_file
-                ));
-
-                full_script
-            } else {
-                tasks_with_placeholder += 1;
-                create_placeholder_script(&task.recipe_name, &task.task_name)
-            }
-        } else {
-            tasks_with_placeholder += 1;
-            create_placeholder_script(&task.recipe_name, &task.task_name)
-        };
-
-        let task_workdir = tmp_dir.join(&task.recipe_name).join(&task.task_name);
-        fs::create_dir_all(&task_workdir)?;
-
-        let output_file = format!("{}.done", task.task_name);
-
-        // Determine network policy based on task type
-        let network_policy = if task.task_name == "do_fetch" || task.task_name.contains("fetch") {
-            NetworkPolicy::FullNetwork  // Real internet access for downloading sources
-        } else {
-            NetworkPolicy::Isolated  // Hermetic builds for everything else
-        };
-
-        let spec = TaskSpec {
-            name: task.task_name.clone(),
-            recipe: task.recipe_name.clone(),
-            script,
-            workdir: task_workdir,
-            env: HashMap::new(),  // TODO: Extract from recipe variables
-            outputs: vec![PathBuf::from(&output_file)],
-            timeout: Some(Duration::from_secs(300)),  // 5 minute default timeout
-            network_policy,
-            resource_limits: ResourceLimits::default(),
-        };
-
-        task_specs.insert(task_key, spec);
-    }
-
-    println!("  Created {} task specifications", task_specs.len());
-    println!("    • {} tasks with real implementations", tasks_with_impl);
-    println!("    • {} tasks with placeholders", tasks_with_placeholder);
-    println!();
-
-    // ========== Step 10: Execute with Interactive Executor ==========
-    println!("🚀 Initializing interactive executor...");
-
-    let cache_dir = build_dir.join("bitzel-cache");
-    fs::create_dir_all(&cache_dir)?;
-
-    let json_output = build_dir.join("execution-report.json");
-    let options = InteractiveOptions {
-        break_on_failure: true,
-        interactive_mode: false,  // Set to true for wave-by-wave execution
-        show_progress: true,
-        export_json: Some(json_output.clone()),
-    };
-
-    let mut executor = InteractiveExecutor::new(&cache_dir, options)?;
-
-    println!("  Cache directory: {}", cache_dir.display());
-    println!("  JSON report:     {}", json_output.display());
-    println!();
-
-    // Execute the task graph
-    println!("⚡ Starting execution...\n");
-
-    match executor.execute_graph(&task_graph, task_specs) {
-        Ok(outputs) => {
-            println!("\n✅ Execution completed successfully!");
-            println!("   Tasks completed: {}", outputs.len());
-
-            // Show statistics
-            let monitor = executor.monitor();
-            let stats = monitor.get_stats();
-            println!("\n📊 Execution Statistics:");
-            println!("   Completed:     {}", stats.completed);
-            println!("   Failed:        {}", stats.failed);
-            println!("   Cached:        {}", stats.cached);
-            println!("   Cache hit:     {:.1}%", stats.cache_hit_rate * 100.0);
-            println!("   Duration:      {:.2}s", stats.total_duration_ms as f64 / 1000.0);
-            println!("   Avg task time: {:.0}ms", stats.avg_task_duration_ms);
-
-            // Export JSON report
-            if json_output.exists() {
-                println!("\n📄 Machine-readable report: {}", json_output.display());
-            }
-        }
-        Err(e) => {
-            eprintln!("\n❌ Execution failed: {}", e);
-
-            // Show partial statistics
-            let monitor = executor.monitor();
-            let stats = monitor.get_stats();
-            eprintln!("\n📊 Partial Statistics:");
-            eprintln!("   Completed:  {}", stats.completed);
-            eprintln!("   Failed:     {}", stats.failed);
-            eprintln!("   Cached:     {}", stats.cached);
-
-            return Err(e.into());
-        }
-    }
-
-    println!("\n💡 Next steps:");
-    println!("  1. Extract actual bash task implementations from recipes");
-    println!("  2. Integrate OverrideResolver for MACHINE/DISTRO conditionals");
-    println!("  3. Apply bbappends using BuildContext.parse_recipe_with_context()");
-    println!("  4. Enable remote caching with Bazel Remote API v2\n");
 
     Ok(())
-}
-
-/// Create a placeholder script for tasks without implementation
-fn create_placeholder_script(recipe_name: &str, task_name: &str) -> String {
-    let output_file = format!("{}.done", task_name);
-    format!(
-        "set -e\necho '[PLACEHOLDER] {} - {}'\nmkdir -p /work/outputs\necho 'completed' > /work/outputs/{}",
-        recipe_name, task_name, output_file
-    )
-}
-
-/// Recursively find all .bb recipe files in a layer
-fn find_recipes(layer: &Path) -> Vec<PathBuf> {
-    walkdir::WalkDir::new(layer)
-        .max_depth(10)
-        .into_iter()
-        .filter_map(|e| e.ok())
-        .filter(|e| e.path().extension().and_then(|s| s.to_str()) == Some("bb"))
-        .map(|e| e.path().to_path_buf())
-        .collect()
 }
